users.js created with auth functions and client.js file renamed and adjusted
